% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{CheckEqns}
\alias{CheckEqns}
\title{This gives important information about a string with equations.}
\usage{
CheckEqns(input_string)
}
\arguments{
\item{input_string}{a string with equations delimited by '\n'}
}
\value{
A list of 4 things are returned. The first is a boolean
which is true if nothing is defined multiple times. The second is
a vector of all the variables used in the equations. The third is
a list of the left hand sides of the equations, and the fourth is a
a list of the right hand sides of the equations.
}
\description{
This gives important information about a string with equations.
}
\examples{
CheckEqns("A = B + C\\nA = C + A")
CheckEqns("A = B*C\\nD = E+A")
}

