% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Transformation.R
\name{Distribute}
\alias{Distribute}
\title{Find all the distinct ways one can take one element from each input set.}
\usage{
Distribute(input_list)
}
\arguments{
\item{input_list}{a list of driver sets/therapeutic sets}
}
\value{
a list containing every possible combination of one element
}
\description{
Find all the distinct ways one can take one element from each input set.
}
\details{
The function removes identical combinations from each input set
  and returns the unique ones such that they are lexicographically sorted
}

