% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Transformation.R
\name{Factor}
\alias{Factor}
\title{Factors a set of driver/therapeutic sets}
\usage{
Factor(input_list)
}
\arguments{
\item{input_list}{a list of driver/therapeutic sets}
}
\value{
list of lists of factors and lists of sets
}
\description{
Factors a set of driver/therapeutic sets
}

