% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{getlastparen}
\alias{getlastparen}
\title{Finds the index of a parenthesis corresponding to the first one.}
\usage{
getlastparen(s)
}
\arguments{
\item{s}{a string}
}
\description{
Finds the index of a parenthesis corresponding to the first one.
}
\examples{
getlastparen("(()())")
}

