% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{InvertList}
\alias{InvertList}
\title{Applies '!' to each element of each driver/therapeutic set}
\usage{
InvertList(input_list)
}
\arguments{
\item{input_list}{list of driver/therapeutic sets}
}
\value{
driver/therapeutic sets with '!' applied to each element
}
\description{
Applies '!' to each element of each driver/therapeutic set
}
\details{
if an element has '!', the '!' is removed
}
\examples{
InvertList(parser("A + !F"))
}

