% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{InvertString}
\alias{InvertString}
\title{Inverts a string by applying '!' to each variable}
\usage{
InvertString(string)
}
\arguments{
\item{string}{}
}
\description{
Inverts a string by applying '!' to each variable
}
\examples{
InvertString("A + B*(C+D)")
}

