% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Transformation Multiply.R
\name{mDisplay}
\alias{mDisplay}
\title{Displays drivers or therapeutics in a simple way}
\usage{
mDisplay(input_list)
}
\arguments{
\item{input_list}{sets of driver/therapeutic sets}
}
\value{
string to diplay
}
\description{
Displays drivers or therapeutics in a simple way
}
\examples{
print(Display(list(list("A",9),list("G",3.567))))
}

