% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser Multiply.R
\name{mparser}
\alias{mparser}
\title{Parses input to interpret it as driver/therapeutic sets}
\usage{
mparser(input)
}
\arguments{
\item{input}{}
}
\value{
the driver/therapeutic sets denoted by the input string
}
\description{
Parses input to interpret it as driver/therapeutic sets
}
\examples{
mparser("(A*(B + C)*D + C*(D+E*(F+C))*A*G)")
}

