% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{Multiply}
\alias{Multiply}
\title{Distributes two lists of lists}
\usage{
Multiply(lon, lo)
}
\arguments{
\item{lon}{a list of driver/therapeutic sets
lo a list of driver/therapeutic sets}
}
\value{
a list of all the ways to combine elements from each list
}
\description{
Distributes two lists of lists
}
\examples{
Multiply(parser("A*B"),parser("C+D"))
}

