% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{parser}
\alias{parser}
\title{Parses input to interpret it as driver/therapeutic sets}
\usage{
parser(input)
}
\arguments{
\item{input}{}
}
\value{
the driver/therapeutic sets denoted by the input string
}
\description{
Parses input to interpret it as driver/therapeutic sets
}
\examples{
parser("(A(B + C)D + C(D+E(F+C))AG)")
}

