% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{Proper}
\alias{Proper}
\title{Checks if a string has the same number of opening and closing parentheses.}
\usage{
Proper(string)
}
\arguments{
\item{string}{}
}
\value{
boolean evaluating the condition
}
\description{
Checks if a string has the same number of opening and closing parentheses.
}
\examples{
Proper("(AB)")
Proper("((((((AB)))))")
}

