% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{RemoveDefine}
\alias{RemoveDefine}
\title{Removes lists in a list containing any defined terms}
\usage{
RemoveDefine(input_list, defined)
}
\arguments{
\item{input_list}{a list of driver/therapeutic sets
defined a list of terms that are to be removed}
}
\value{
driver/therapeutic sets without the defined terms
}
\description{
Removes lists in a list containing any defined terms
}
\examples{
RemoveDefine(parser("A*B+A*C+D*B*+E*F"),list("A"))
}

