% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Transformation.R
\name{RemoveParen}
\alias{RemoveParen}
\title{Removes some unnecessary parentheses from expressions}
\usage{
RemoveParen(string)
}
\arguments{
\item{string}{}
}
\value{
string without some parentheses that doesn't change the meaning of the expression
}
\description{
Removes some unnecessary parentheses from expressions
}
\examples{
RemoveParen("(AB)")
}

