% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{SatisfiableList}
\alias{SatisfiableList}
\title{Removes non-satisfiable elements of a list}
\usage{
SatisfiableList(input_list)
}
\arguments{
\item{input_list}{list of driver/therapeutic sets}
}
\value{
satisfiable driver/therapeutic sets
}
\description{
Removes non-satisfiable elements of a list
}
\details{
any driver/therapeutic set with some string
as well as the string with '!' is not satisfiable and is removed
}
\examples{
SatisfiableList(parser("A&!A+B"))
}

