% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Transformation.R
\name{Simplify}
\alias{Simplify}
\title{Finds minimal sets. Given an input, this function returns all the distinct
sets that are not proper subsets of other sets.}
\usage{
Simplify(input_list)
}
\arguments{
\item{input_list}{a list of driver/therapeutic sets}
}
\value{
minimal sets from input
}
\description{
Finds minimal sets. Given an input, this function returns all the distinct
sets that are not proper subsets of other sets.
}

