% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Transformation.R
\name{Sort}
\alias{Sort}
\title{Sorts a list}
\usage{
Sort(list)
}
\arguments{
\item{list}{}
}
\value{
sorted list. If the input list has numbers, they will be
  numerically sorted. If strings are inputed, they will be lexicographically sorted
}
\description{
Sorts a list
}

