% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{Sub}
\alias{Sub}
\title{Substitutes equations into expr n number of times.}
\usage{
Sub(expr, n, left, right)
}
\arguments{
\item{expr}{expression to be substituted
n number of times to substitute
left left hand sides of equations
right right hand sides of equations}
}
\value{
substituted equations
}
\description{
Substitutes equations into expr n number of times.
}
\examples{
Sub("A",3,"A","(A+B)")
}

