% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Transformation.R
\name{Transform}
\alias{Transform}
\title{Switches between driver sets and therapeutic sets.
  Does so by distributing the sets and simplifying}
\usage{
Transform(input_list)
}
\arguments{
\item{input_list}{a list of all driver/therapeutic sets}
}
\value{
the corresponding complement set
}
\description{
Switches between driver sets and therapeutic sets.
  Does so by distributing the sets and simplifying
}

