% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Transformation.R
\name{Factor}
\alias{Factor}
\title{Factors a set of driver/therapeutic sets}
\usage{
Factor(input_list)
}
\arguments{
\item{input_list}{a list of driver/therapeutic sets}
}
\value{
String of factored expression
}
\description{
Factors a set of driver/therapeutic sets
}
\examples{
Factor(parser("AB + AC + AD"))
Factor(parser("BD + AC + AD + BC"))
}

